extract legacy statistics module
